==> Scripts/s03_summary_excluded_data.Rmd <==
---
title: "Summary of excluded and commented data"
author: "A. Analyst"
date: "2018-02-19"
output: html_document
params:
  print_results: FALSE
---
```{r setup, include=FALSE}
###################################################
# s03_summaryOfExcludedData.R
# Description: Summary of data excluded with C="C
# Dependencies: s01_datasetPrep.R
###################################################
library(rprojroot)
knitr::opts_chunk$set(echo=F)
knitr::opts_knit$set(root.dir=find_root(has_file("OpenProject.Rproj"))) 
# knit in this directory -> everything after this chunk is relative top level dir

# Note: R markdown opens a new R session, your global environment is not available.
```

Load the dataset as prepared by "s01_datasetPrep.R":
```{r, echo=T, message=F}
load(file=file.path("./Scripts","s01.RData"))
# load packages
source(file=file.path(setup_dir,"setup01_projectPackages.R"))
```

```{r}
# The outputs from this scripts are: 
# 1. a summary table of number of excluded data and reasons (split by >/< LLOQ)
# 2. a table listing each excluded record
# Both to be used in the report. 
```

# Summary of excluded subjects, concentrations and doses
Number of subjects excluded and total number in final data:
```{r}
nTotal <- length(unique(rawdata$OSID))
nIncl <- length(unique(data$OSID))
print(paste("Excluded: ", nIncl-nTotal))
print(paste("Included: ", nIncl))
```

Total number of excluded rows: 
```{r}
excluded <- rawdata %>% filter(C=="C")
nrow(excluded)

excl_doses <- excluded %>% filter(EVID==1)
excl_conc <- excluded %>% filter(EVID==0)
```
Out of which `r nrow(excl_doses)` rows are doses and `r nrow(excl_conc)` concentrations.

Is anything else than EVID=1 or 0 that is excluded?
```{r}
if(! sum(nrow(excl_doses), nrow(excl_conc)) ==  nrow(excluded)){
  print("Yes")
} else {
 print("No") 
}
```

How much of the excluded concentration data is missing or BLQ?
```{r}
excl_conc_missing <- excl_conc %>% filter(MDV==1)
paste0(nrow(excl_conc_missing)," (", round(100*(nrow(excl_conc_missing) / nrow(excl_conc))), "%)")
```

# Review and summary of comments
## Excluded doses
```{r}
unique(excl_doses[,"COMMENT"])
```

## Excluded concentrations
```{r}
unique(excl_conc[,"COMMENT"])
```
### Reason for exclusion split by BLQ/non-BLQ:
```{r}
# In this dataset there are not multiple comments in one row/cell. 
# If that is the case, you need to be careful not to count it twice/overwrite the reason.

excl_conc$REASON[str_detect(excl_conc$COMMENT, "Pre first dose sample") & excl_conc$BLQ==1] <- 
  "Pre first dose sample, BLQ"
excl_conc$REASON[str_detect(excl_conc$COMMENT, "Pre first dose sample") & excl_conc$BLQ==0] <- 
  "Pre first dose sample, Non-BLQ"

excl_conc$REASON[str_detect(excl_conc$COMMENT, "Randomly selected for illustration") & excl_conc$BLQ==1] <- 
  "Random exclusion, BLQ"
excl_conc$REASON[str_detect(excl_conc$COMMENT, "Randomly selected for illustration") & excl_conc$BLQ==0] <-
  "Random exclusionm, Non-BLQ"

total_conc <- rawdata %>% filter(EVID==0)

summary <- 
  excl_conc %>%
  group_by(REASON) %>% 
  summarize(n = n(), 
            percentOfTotal = round(100 * (n / nrow(total_conc)), digits=1))
summary
```


```{r, include=F}
if(params$print_results){
  # Output in reportable table
  outTab <- summary %>% 
    mutate(Values = paste0(n, " (", percentOfTotal, "%)")) %>% 
    select(REASON, Values)
  
  write.csv(outTab,
            file = file.path(res_eda_dir, "omittedConcDataSummary.csv"),
            row.names = F)
}
```

### How much of the non-BLQ data are sparse samples: 
```{r}
# Note, sparse data is those without a OCC assignment in this particular dataset
# Hence code might need to be changed depending on current data
sparse <- 
  excl_conc %>%
  filter(is.na(OCC)) %>% 
  group_by(REASON) %>% 
  summarize(n = n(), 
            percentOfTotal = round(100 * (n / nrow(total_conc)), digits=1))
sparse
```


# Review of comments not leading to exclusion
```{r}
# Use the dataset 'data' for this since all C=="C" is commented out
unique(data$COMMENT)

data %>%
  filter(! (COMMENT %in% c(" ", "") | is.na(COMMENT))) %>% 
  group_by(EVID, COMMENT) %>% 
  summarize(n = n(), 
            percentOfTotal = round(100 * (n / nrow(data)), digits=2))
```
