==> Scripts/s05_EDA_conc_time.Rmd <==
---
title: "Exploratory data analysis: Concentrations versus time"
author: "A. Analyst"
date: "2018-02-20"
output: 
  html_document:
    toc: true
    toc_depth: 2
    number_sections: true
params:
  print_results: FALSE
---
```{r setup, include=F}
###################################################
# s05_EDAConcTime.Rmd
# Description: Exploratory data analysis of concentrations vs time for population PK analysis
# Dependencies: s01_datasetPrep.R / s01.RData
###################################################
library(rprojroot)
knitr::opts_chunk$set(echo=F)
knitr::opts_knit$set(root.dir=find_root(has_file("OpenProject.Rproj"))) 
# for correct printing of dust tables
options(pixie_interactive=F)
# Note: R markdown opens a new R session, your global environment is not available.
```

This script uses the data.frame "data", loaded from s01.RData. That is, all rows with C=="C" has been excluded for exploratory data analysis.
```{r environmentPrep, echo=T, message=F}
# ------------------------------------------------------------------
#  Prepare environment
# ------------------------------------------------------------------
# load packages
source(file=file.path("Scripts","Setup","setup01_projectPackages.R")) 
# load output from s01_DatasetPrep.R
load(file=file.path("Scripts","s01.RData"))

# White background in plots
theme_set(theme_bw()) # to be replaced with a azTheme
update_geom_defaults("point", list(shape = 1))
```

Are the plots and tables also being written to file?
```{r}
params$print_results
```

# Numeric summaries
The total number of observations in the dataset (with/without BLQ data) and the %BLQ.
```{r concentrationSummaries}
# Total number of concentrations in dataset
nrow(conc_data)          # All
nrow(conc_data_noBLQ)    # >BLQ
round((nrow(conc_data) - nrow(conc_data_noBLQ))/nrow(conc_data)*100, digits = 2) # percent BLQ
```

Number of samples and occasions/subject (with/without BLQ)
```{r concentrationSummariesBySubject}
conc_per_pat <- # total
  conc_data %>%
  group_by(OSID, OSTUDYID) %>% 
  summarize(nDV=length(DV), 
            nOcc = ifelse(all(is.na(unique(OCC))), 0, 
                          max(as.numeric(as.character(OCC)), na.rm=T)),
            maxTAFD = max(TAFD, na.rm=T))
summary(conc_per_pat[,c("nDV","nOcc","maxTAFD")])

conc_per_pat_noBLQ <-  # >BLQ
  conc_data_noBLQ %>% 
  group_by(OSID, OSTUDYID) %>% 
  summarize(nDV=length(DV), 
            nOcc = ifelse(all(is.na(unique(OCC))), 0, 
                          max(as.numeric(as.character(OCC)), na.rm=T)),
            maxTAFD = max(TAFD, na.rm=T))
summary(conc_per_pat_noBLQ[,c("nDV","nOcc","maxTAFD")])
```

Number of individuals and samples per dose group
```{r samples, warning=F}
conc_per_dose <- # total
  conc_data %>% 
  group_by(DOSE) %>% 
  summarize(nSubjects = length(unique(OSID)),
            nConc=length(DV)) %>% 
  mutate(concPercent = 
           round(100*(nConc/nrow(conc_data)), digits=1))

conc_per_dose_noBLQ <-  # >BLQ
  conc_data_noBLQ %>% 
  group_by(DOSE) %>% 
  summarize(nSubjects = length(unique(OSID)),
            nConc=length(DV)) %>% 
  mutate(concPercent = 
           round(100*(nConc/nrow(conc_data_noBLQ)), digits=1))

# % BQL per dose group
conc_per_dose <- 
  conc_per_dose %>% 
  mutate(percentBlqByDose = 
           signif(100-(conc_per_dose_noBLQ$nConc/nConc)*100, digits=3))

## merge to reportable table
conc_per_dose_tab <- 
  conc_per_dose %>% 
  mutate(nConc2 = paste0(nConc," (", concPercent,")")) %>% 
  select(DOSE, nSubjects, nConc2, percentBlqByDose) %>% 
  rename(nConc = nConc2, 
         Dose = DOSE)

dust(conc_per_dose_tab, 
     caption="Number of individuals and samples per dose group")
```

```{r, include=F}
# --------- Save to file
if(params$print_results){
  # Note: if subjects are gived multiple doses they are counted several times
  write.csv(conc_per_dose_tab, 
            file = file.path(res_eda_dir, "concentrations_by_dose.csv"), 
            row.names = F)
}
```

# Concentrations versus time, stratified by dose and study
## Concentrations vs time after first dose, stratified by study and dose group
Lines connect data from one occasion within a subject. Colour indicate subject. Points indicate measured data. Dashed blue line show the lower limit of quantification. BLQ data is prited as LLOQ/2. 

```{r ConcTime, fig.height=6.5, fig.width=8.5, warning=F, message=F}
# Shortest and longest TAFD to set the same x-axes in all plot
tafd_range <- conc_data %>% 
  summarize(min = min(TAFD), 
            max = max(TAFD))

# List of plots
plots <- vector("list", length(conc_data_studydose_split))
for(i in 1:length(conc_data_studydose_split)){
  p <-
    arrangeGrob(
      # linear
      gg_conc_time(conc_data_studydose_split[[i]], x=TAFD, 
                 y=DV, color=OSID, occ=OCC) + 
        facet_wrap(~STUDYDOSESPLIT) + 
        coord_cartesian(xlim=c(0, tafd_range$max)) +
        guides(colour="none", shape="none") + 
        labs(x=labs_TAFD, y=labs_conc),
      # log
      gg_conc_time(conc_data_studydose_split[[i]], x=TAFD, 
                 y=DV, color=OSID, occ=OCC) + 
        facet_wrap(~STUDYDOSESPLIT) + 
        coord_cartesian(xlim=c(0, tafd_range$max)) +
        guides(colour="none", shape="none") + 
        labs(x=labs_TAFD, y=labs_conc) + 
        scale_y_log10(), 
      nrow=2)
  plots[[i]] <- p
}
# print each on a separate page
invisible(lapply(plots, grob_draw))

if(params$print_results){
  pdf(file=file.path(res_eda_dir, paste0("conc_vs_TAFD_by_study_dose_", delivery_date,".pdf")),
      height=6.5, width=8.5)
  invisible(lapply(plots, grob_draw))
  dev.off()
}
``` 

## Concentrations versus time after dose, stratified by study and dose
*Right*: un-transformed scale. *Left*: log-transformed y-axis. Lines connect data from one occasion within a subject. Colour indicate subject. Points indicate measured data. Dashed blue line show the lower limit of quantification. BLQ data is prited as LLOQ/2. 

```{r ConcTad, fig.height=3, fig.width=8.5, warning=F, message=F}
# use max TAPD to set the same x axis on all graphs
tapd_range <- conc_data %>% 
  summarize(min = min(TAPD), 
            max = max(TAPD))

# List of plots
plots <- vector("list", length(conc_data_studydose_split))
for(i in 1:length(conc_data_studydose_split)){
  p <-
    arrangeGrob(
      # linear
      gg_conc_time(conc_data_studydose_split[[i]], x=TAPD, 
                 y=DV, color=OSID, occ=OCC) + 
        facet_wrap(~STUDYDOSESPLIT) + 
        coord_cartesian(xlim=c(0, tapd_range$max)) +
        guides(colour="none", shape="none") + 
        labs(x=labs_TAPD, y=labs_conc),
      # log
      gg_conc_time(conc_data_studydose_split[[i]], x=TAPD, 
                 y=DV, color=OSID, occ=OCC) + 
        facet_wrap(~STUDYDOSESPLIT) + 
        coord_cartesian(xlim=c(0, tapd_range$max)) +
        guides(colour="none", shape="none") + 
        labs(x=labs_TAPD, y=labs_conc) + 
        scale_y_log10(), 
      nrow=1)
  plots[[i]] <- p
}
# print each on a separate page
invisible(lapply(plots, grob_draw))

if(params$print_results){
  pdf(file=file.path(res_eda_dir,
                     paste0("conc_vs_TAPD_by_study_dose_", delivery_date,".pdf")),
      height=3, width=8.5)
  invisible(lapply(plots, grob_draw))
  dev.off()
}
```

## Concentrations versus time after dose, stratified by study and dose (zoom)
### First 12 hours
*Right*: un-transformed scale. *Left*: log-transformed y-axis. Lines connect data from one occasion within a subject. Colour indicate subject. Points indicate measured data. Dashed blue line show the lower limit of quantification. BLQ data is prited as LLOQ/2. 

```{r ConcTadZoom, fig.height=3, fig.width=8.5, warning=F, message=F} 
# If you want to zoom in on a specific time scale e.g. absorption phase or first 12/24 hr 
x_max <- 12

# List of plots
plots <- vector("list", length(conc_data_studydose_split))
for(i in 1:length(conc_data_studydose_split)){
  p <-
    arrangeGrob(
      # linear
      gg_conc_time(conc_data_studydose_split[[i]], x=TAPD, 
                 y=DV, color=OSID, occ=OCC) + 
        facet_wrap(~STUDYDOSESPLIT) + 
        coord_cartesian(xlim=c(0, x_max)) +
        guides(colour="none", shape="none") + 
        labs(x=labs_TAPD, y=labs_conc),
      # log
      gg_conc_time(conc_data_studydose_split[[i]], x=TAPD, 
                 y=DV, color=OSID, occ=OCC) + 
        facet_wrap(~STUDYDOSESPLIT) + 
        coord_cartesian(xlim=c(0, x_max)) +
        guides(colour="none", shape="none") + 
        labs(x=labs_TAPD, y=labs_conc) + 
        scale_y_log10(), 
      nrow=1)
  plots[[i]] <- p
}
# print each on a separate page
invisible(lapply(plots, grob_draw))

if(params$print_results){
  pdf(file=file.path(res_eda_dir,
                     paste0("conc_vs_TAPD_by_study_dose_", x_max,
                            "_", delivery_date,".pdf")),
      height=3, width=8.5)
  invisible(lapply(plots, grob_draw))
  dev.off()
}
```

### Absorption phase: first 4 hours
*Right*: un-transformed scale. *Left*: log-transformed y-axis. Lines connect data from one occasion within a subject. Colour indicate subject. Points indicate measured data. Dashed blue line show the lower limit of quantification. BLQ data is prited at LLOQ/2. 

```{r ConcTadAbs, fig.height=3, fig.width=8.5, warning=F, message=F} 
# Zoom in on absorption phase
x_max <- 4

# List of plots
plots <- vector("list", length(conc_data_studydose_split))
for(i in 1:length(conc_data_studydose_split)){
  p <-
    arrangeGrob(
      # linear
      gg_conc_time(conc_data_studydose_split[[i]], x=TAPD, 
                 y=DV, color=OSID, occ=OCC) + 
        facet_wrap(~STUDYDOSESPLIT) + 
        coord_cartesian(xlim=c(0, x_max)) +
        guides(colour="none", shape="none") + 
        labs(x=labs_TAPD, y=labs_conc),
      # log
      gg_conc_time(conc_data_studydose_split[[i]], x=TAPD, 
                 y=DV, color=OSID, occ=OCC) + 
        facet_wrap(~STUDYDOSESPLIT) + 
        coord_cartesian(xlim=c(0, x_max)) +
        guides(colour="none", shape="none") + 
        labs(x=labs_TAPD, y=labs_conc) + 
        scale_y_log10(), 
      nrow=1)
  plots[[i]] <- p
}
# print each on a separate page
invisible(lapply(plots, grob_draw))

if(params$print_results){
  pdf(file=file.path(res_eda_dir, 
                     paste0("absorption_phase_", x_max,"_", delivery_date,".pdf")),
      height=3, width=8.5)
  invisible(lapply(plots, grob_draw))
  dev.off()
}
```

## Comparison of single and multiple dose occasions
*Right*: un-transformed scale. *Left*: log-transformed y-axis. Lines connect data from one occasion within a subject. Colour indicate subject. Points indicate measured data. Dashed blue line show the lower limit of quantification. BLQ data is prited at LLOQ/2. 

```{r SingleVsMultiple, fig.height=6, fig.width=8.5, warning=F, message=F} 
# First 24 hours after dose, compare single and steady state
# (Requires that DOSEFLAG has been added to the dataset in s01_dataset_preparation.R)
x_max <- 24

# List of plots
plots <- vector("list", length(conc_data_studydose_split))
for(i in 1:length(conc_data_studydose_split)){
  # remove sparse occasions (DOSEFLAG = NA)
  dat <- conc_data_studydose_split[[i]] %>% 
    filter(!is.na(DOSEFLAG))
  
  p <-
    arrangeGrob(
      # linear
      gg_conc_time(dat, x=TAPD,
                 y=DV, color=OSID, occ=OCC) + 
        facet_wrap(~STUDYDOSESPLIT+DOSEFLAG, nrow = 2) + 
        coord_cartesian(xlim=c(0, x_max)) +
        guides(colour="none", shape="none") + 
        labs(x=labs_TAPD, y=labs_conc),
      # log
      gg_conc_time(dat, x=TAPD, 
                 y=DV, color=OSID, occ=OCC) +
        facet_wrap(~STUDYDOSESPLIT+DOSEFLAG, nrow = 2) + 
        coord_cartesian(xlim=c(0, x_max)) +
        guides(colour="none", shape="none") + 
        labs(x=labs_TAPD, y=labs_conc) + 
        scale_y_log10(), 
      nrow=1)
  plots[[i]] <- p
}
# print each on a separate page
invisible(lapply(plots, grob_draw))

if(params$print_results){ 
  pdf(file=file.path(res_eda_dir,
                     paste0("single_vs_multiple_dose_", x_max,"_", delivery_date,".pdf")),
      height=6, width=8.5)
  invisible(lapply(plots, grob_draw))
  dev.off()
}
```

# Individual plots of concentrations vs time
The section below generates a list of individual plots with different combinations of the following settings:

* Concentration versus time after first dose *or* time after dose

* Log scale and untransformed scale

* Free/fixed scales for axes across individuals

Not all plots are going to be useful, you can just comment out/delete the ones you do not want to keep.

```{r individual, fig.height=8.5, fig.width=11, message=F, warning=F} 
range_axes <- conc_data %>% 
  summarize(minDV = min(DV, na.rm = T),
            maxDV = max(DV, na.rm = T), 
            maxTAFD = max(TAFD, na.rm = T),
            maxTAPD = max(TAPD, na.rm = T))

# 1. Conc vs TAFD. Same axes across subjects. 
individualPlots <- list(
  gg_title_plot("Concentration vs. time after first dose \n Same axes across subjects"))

p1 <- vector("list", length(conc_data_id_splits))
for(i in 1:(length(conc_data_id_splits))){
  p <-
    gg_conc_time(conc_data_id_splits[[i]], x=TAFD, y=DV, 
               color=REGIMEN, occ=OCC) + 
    facet_wrap(~OSID+COHORT, labeller="label_both",
               nrow=3, ncol=4) +
    # set scales for axes
    coord_cartesian(ylim=c(0, range_axes$maxDV), 
                    xlim=c(0, range_axes$maxTAFD)) +
    guides(shape="none") + 
    labs(x=labs_TAFD, y=labs_conc)
  p1[[i]] <- p
}
individualPlots <- c(individualPlots, p1)

# 2. Log Conc vs. TAFD. Same axes across subjects
individualPlots <- 
  c(individualPlots, list(
    gg_title_plot("Concentrations vs. time after first dose \n Same axes across subjects \n\n Semi-log")))

p1 <- vector("list", length(conc_data_id_splits))
for(i in 1:(length(conc_data_id_splits))){
  p <-
    gg_conc_time(conc_data_id_splits[[i]], x=TAFD, y=DV, 
               color=REGIMEN, occ=OCC) + 
    facet_wrap(~OSID+COHORT, labeller="label_both", 
               nrow=3, ncol=4) +
    # set scales for axes
    coord_cartesian(ylim=c(range_axes$minDV, range_axes$maxDV), 
                    xlim=c(0, range_axes$maxTAFD)) +
    scale_y_log10(breaks=c(1,10,100,1000)) +
    guides(shape="none") + 
    labs(x=labs_TAFD, y=labs_conc)
  p1[[i]] <- p
}
individualPlots <- c(individualPlots, p1)

# 3. Conc vs TAFD. Free scales on both axes
individualPlots <-
  c(individualPlots, list(
    gg_title_plot("Concentration. vs time after first dose \n Free scales on axes")))

p1 <- vector("list", length(conc_data_id_splits))
for(i in 1:(length(conc_data_id_splits))){
  p <-
    gg_conc_time(conc_data_id_splits[[i]], x=TAFD, y=DV, 
               color=REGIMEN, occ=OCC) + 
    facet_wrap(~OSID+COHORT, labeller="label_both", 
               nrow=3, ncol=4, 
               scales="free") +
    guides(shape="none") + 
    labs(x=labs_TAFD, y=labs_conc)
  p1[[i]] <- p
}
individualPlots <- c(individualPlots, p1)

# 4. Log Conc vs TAFD. Free scales on axes
individualPlots <-
  c(individualPlots,list(
    gg_title_plot("Concentration vs. time after first dose \n Free scales on axes \n\n Semi-log")))

p1 <- vector("list", length(conc_data_id_splits))
for(i in 1:(length(conc_data_id_splits))){
  p <-
    gg_conc_time(conc_data_id_splits[[i]], x=TAFD, y=DV, 
               color=REGIMEN, occ=OCC) + 
    facet_wrap(~OSID+COHORT, labeller="label_both", 
               nrow=3, ncol=4, 
               scales="free") +
    scale_y_log10(breaks=c(1,10,100,1000)) +
    guides(shape="none") + 
    labs(x=labs_TAFD, y=labs_conc)
  p1[[i]] <- p
}
individualPlots <- c(individualPlots, p1)

# 5. Conc vs. TAPD. Same y-axis across subjects. 
individualPlots <- 
  c(individualPlots, list(
    gg_title_plot("Concentration vs. time after dose \n Same axes across subjects")))

p1 <- vector("list", length(conc_data_id_splits))
for(i in 1:(length(conc_data_id_splits))){
  p <-
    gg_conc_time(conc_data_id_splits[[i]], x=TAPD, y=DV, 
               color=REGIMEN, occ=OCC) + 
    facet_wrap(~OSID+COHORT, labeller="label_both",
               nrow=3, ncol=4) +
    # set scales for axes
    coord_cartesian(ylim=c(0, range_axes$maxDV), 
                    xlim=c(0, range_axes$maxTAPD)) +
    guides(shape="none") + 
    labs(x=labs_TAPD, y=labs_conc)
  p1[[i]] <- p
}
individualPlots <- c(individualPlots, p1)

# 6. Log Conc vs. TAPD. Same y-axis across subjects
individualPlots <-
  c(individualPlots,list(
    gg_title_plot("Concentration vs. time after dose \n Same axes across subjects \n\n Semi-log")))

p1 <- vector("list", length(conc_data_id_splits))
for(i in 1:(length(conc_data_id_splits))){
  p <-
    gg_conc_time(conc_data_id_splits[[i]], x=TAPD, y=DV, 
               color=REGIMEN, occ=OCC) + 
    facet_wrap(~OSID+COHORT, labeller="label_both",
               nrow=3, ncol=4) +
    # set scales for axes
    coord_cartesian(ylim=c(range_axes$minDV, range_axes$maxDV),
                    xlim=c(0, range_axes$maxTAPD)) +
    scale_y_log10(breaks=c(1,10,100,1000)) +
    guides(shape="none") + 
    labs(x=labs_TAPD, y=labs_conc)
  p1[[i]] <- p
}
individualPlots <- c(individualPlots, p1)

# 7. Conc vs TAPD - Free scales
individualPlots <-
  c(individualPlots,list(
    gg_title_plot("Concentration vs. time after dose \n Free scales on axes")))

p1 <- vector("list", length(conc_data_id_splits))
for(i in 1:(length(conc_data_id_splits))){
  p <-
    gg_conc_time(conc_data_id_splits[[i]], x=TAPD, y=DV, 
               color=REGIMEN, occ=OCC) + 
    facet_wrap(~OSID+COHORT, labeller="label_both",
               nrow=3, ncol=4, 
               scales="free") +
    guides(shape="none") + 
    labs(x=labs_TAPD, y=labs_conc)
  p1[[i]] <- p
}
individualPlots <- c(individualPlots, p1)

## 8. Log conc vs TAPD - Free scales
individualPlots <-
  c(individualPlots,list(
    gg_title_plot("Concentration vs. time after dose \n Free scales on axes \n Semi-log")))

p1 <- vector("list", length(conc_data_id_splits))
for(i in 1:(length(conc_data_id_splits))){
  p <-
    gg_conc_time(conc_data_id_splits[[i]], x=TAPD, y=DV, 
               color=REGIMEN, occ=OCC) + 
    facet_wrap(~OSID+COHORT, labeller="label_both",
               nrow=3, ncol=4, 
               scales="free") +
    scale_y_log10(breaks=c(1,10,100,1000)) +
    guides(shape="none") + 
    labs(x=labs_TAPD, y=labs_conc)
  p1[[i]] <- p
}
individualPlots <- c(individualPlots, p1)

# 9. Log Conc vs. TAPD. Same y-axis across subjects, TAD <12 h
x_max <- 12

individualPlots <-
  c(individualPlots,list(
    gg_title_plot(paste("Concentration vs. time after dose \n TAD <", x_max ,"\n\n Semi-log"))))

p1 <- vector("list", length(conc_data_id_splits))
for(i in 1:(length(conc_data_id_splits))){
  p <-
    gg_conc_time(conc_data_id_splits[[i]], x=TAPD, y=DV, 
               color=REGIMEN, occ=OCC) + 
    facet_wrap(~OSID+COHORT, labeller="label_both",
               nrow=3, ncol=4) +
    # set scales for axes
    coord_cartesian(ylim=c(range_axes$minDV, range_axes$maxDV), 
                    xlim=c(0, x_max)) +
    scale_y_log10(breaks=c(1,10,100,1000)) +
    guides(shape="none") + 
    labs(x=labs_TAPD, y=labs_conc)
  p1[[i]] <- p
}
individualPlots <- c(individualPlots, p1)

## This list takes quite some time to print...
invisible(lapply(individualPlots, grob_draw))

if(params$print_results){
  pdf(file=file.path(res_eda_dir, 
                     paste0("individual_conc_vs_time_", delivery_date,".pdf")),
      height=8.5, width=11)
  invisible(lapply(individualPlots, grob_draw))
  dev.off()
}
```
