==> Scripts/s06_EDA_PK_linearity.Rmd <==
---
title: "Exploratory data analysis: PK Linearity"
author: "A. Analyst"
date: "2018-02-20"
output: 
  html_document:
    toc: true
    toc_depth: 2
    number_sections: true
params:
  print_results: FALSE
---
```{r setup, include=F}
###################################################
# s06_EDAPkLinearity.Rmd
# Description: Exploratory data analysis of PK linearity for population PK analysis
# Dependencies: s01_datasetPrep.R
###################################################
library(rprojroot)
knitr::opts_chunk$set(echo=F)
knitr::opts_knit$set(root.dir=find_root(has_file("OpenProject.Rproj"))) 
# for correct printing of dust tables
options(pixie_interactive=F)

# Note: R markdown opens a new R session, your global environment is not available.
```

This script uses the data.frame "data", loaded from s01.RData. That is, all rows with C=="C" has been excluded for exploratory data analysis.
```{r environmentPrep, echo=T, message=F}
# ------------------------------------------------------------------
#  Prepare environment
# ------------------------------------------------------------------
# load packages
source(file=file.path("Scripts","Setup","setup01_projectPackages.R")) 
# load output from s01_DatasetPrep.R
load(file=file.path("Scripts","s01.RData"))

# White background in plots
theme_set(theme_bw()) # to be replaced with a azTheme
update_geom_defaults("point", list(shape = 1))
```

Are the plots and tables also being written to file?
```{r}
params$print_results
```

# Included data
```{r, include=F} 
# The sections below will need to be tweaked for each specific project. 
# This should only serve as an example

# ------------------------------------------------------------------
#  Settings for NCA:
# ------------------------------------------------------------------
nca_intervals <-
  data.frame(start     = 0,
             end       = c(24),
             cmax      = T,
             tmax      = T,
             aucinf.obs= T,
             auclast   = T,
             tlast     = T,
             tfirst    = T,
             half.life = T,
             lambda.z  = T,
             lambda.z.n.points = T)

# ------------------------------------------------------------------
#  Extract single dose data
# ------------------------------------------------------------------
nca_single <-  data %>% 
  filter(!is.na(DOSEFLAG) & DOSEFLAG =="Single dose")

# Use ID.OCC as ID to include multiple single doses from the same patient
# ID.OCC are missing for doses => fill backwards
nca_single <- nca_single %>% 
  mutate(ID.OCC = na.locf(ID.OCC, fromLast = T), 
         ID.OCC = as.numeric(as.character(ID.OCC)))

# Dose dataset
nca_single_doses <- nca_single %>% 
  filter(EVID==1) # dosing events
# unique(nca_single_doses$AMT)

## Concentration dataset
nca_single_concs <- nca_single %>% 
  filter(!is.na(BLQ) & BLQ=="Non-BLQ" & !is.na(DV)) # remove all BLQs (as well as dosing events)

## You may have to remove other missing concentrations (MDV==1)
# nca_single_concs <- nca_single_concs %>% 
#   filter(!(!is.na(MDV) & MDV==1 & is.na(BLQ)))

# Add conc=0 at time 0 (or whatever you think is appropriate)
nca_single_conc_0 <- 
  data.frame(ID.OCC = unique(nca_single_concs$ID.OCC),
             TAFD=0, 
             TAPD=0, 
             DV=0)

# join and sort
nca_single_concs <- 
  full_join(nca_single_concs, nca_single_conc_0) %>% 
  arrange(ID.OCC, TAFD)
```

## Plots of occasions included for single dose NCA parameters
Lines connect data from one occasion, points show measured concentrations, the blue horizontal line indicated time of dosing. 

```{r, fig.height=8.5, fig.width=11, warning=F, message=F}
# ------------------------------------------------------------------
#  Plot single dose data
# ------------------------------------------------------------------
# split by id.occ
nca_single_dose_split <- ind_data_split(nca_single_doses, 
                                        n_per_page = 12, id="ID.OCC")
nca_single_concs_split <- ind_data_split(nca_single_concs,
                                         n_per_page = 12, id="ID.OCC")

# IndividualPK profiles for single dose
p1 <- vector("list", length(nca_single_concs_split))
for(i in 1:length(nca_single_concs_split)){
  p <-
    # concentrations
    ggplot(data=nca_single_concs_split[[i]], aes(x=TAPD, y=DV)) +
    geom_line() + geom_point() +
    # doses
    geom_vline(data = nca_single_dose_split[[i]], aes(xintercept=TAPD),
               col="blue", linetype="dashed") +
    facet_wrap(~ID.OCC, nrow=3, ncol=4,
               scales="free_y", labeller="label_both") +
    labs(x=labs_TAPD, y=labs_conc)
  p1[[i]] <- p
}

invisible(lapply(p1, grob_draw))
```

```{r, include=F}
if(params$print_results){
  pdf(file=file.path(res_eda_dir, paste0("nca_single_dose_", delivery_date,".pdf")),
      height=8.5, width=11)
  invisible(lapply(p1, grob_draw))
  dev.off()
}
```

```{r, include=F}
# ------------------------------------------------------------------
# Merge conc and dose data into NCA object
# ------------------------------------------------------------------
nca_single_data <-
  PKNCAdata(nca_single_concs, nca_single_doses,
            formula.conc = DV~TAPD | ID.OCC,
            formula.dose = AMT~TAPD | ID.OCC,
            intervals = nca_intervals)

# ------------------------------------------------------------------
# Run NCA, extract and re-format results
# ------------------------------------------------------------------
nca_results_single <- pk.nca(nca_single_data)

# nca results
nca_results_single_extract <- nca_results_single$result
nca_results_single_extract <- 
  spread(nca_results_single_extract[, c("ID.OCC","PPTESTCD","PPORRES")],
         key = PPTESTCD, value = PPORRES)

# Merge back with covariate data (extract from dose object)
cov_extract_single <- 
  nca_results_single$data$dose$data[, c("ID.OCC", cols_cont_cov, base_cont_cov,
                                        cols_cat_cov, base_cat_cov,
                                      "OSID","OSTUDYID","AMT","FREQ", "DOSEFLAG")]

# unique(unique(covExtractSingle$ID.OCC) == unique(nca_results_single_extract$ID.OCC))
results_single <- 
  full_join(nca_results_single_extract, cov_extract_single) %>% 
  rename(dose = AMT)
```

```{r, include=F}
# ------------------------------------------------------------------
#  Extract multiple dose/steady state data
# ------------------------------------------------------------------
nca_multiple <- data %>% 
  filter(!is.na(DOSEFLAG) & DOSEFLAG =="Multiple dose")

# Use ID.OCC as ID to include multiple occasions from the same patient
# ID.OCC are missing for doses => fill backwards
nca_multiple <- nca_multiple %>% 
  mutate(ID.OCC = na.locf(ID.OCC, fromLast = T), 
         ID.OCC = as.numeric(as.character(ID.OCC)))

# Dose dataset
nca_multiple_doses <- nca_multiple %>% 
  filter(EVID==1) # dosing events

## Concentration dataset
nca_multiple_concs <- nca_multiple %>% 
  filter(!is.na(BLQ) & BLQ=="Non-BLQ" & !is.na(DV)) # remove all BLQs (also removes doses)

## You may have to remove other missing concentrations (MDV==1)
# nca_multiple_conc <- 
#   nca_multiple_conc %>% 
#   filter(!(!is.na(MDV) & MDV==1 & is.na(BLQ)))

# Assume no accumulation and add a 0 conc at TAPD=0
# (or whatever you think is appropriate in your case)
nca_multiple_conc_0 <- 
  data.frame(ID.OCC = unique(nca_multiple_concs$ID.OCC),
             TAFD=0, 
             TAPD=0, 
             DV=0)

# join and sort
nca_multiple_concs <- full_join(nca_multiple_concs, nca_multiple_conc_0) %>% 
  arrange(ID.OCC, TAFD)

```

## Plots of occasions included for multiple dose/steady state NCA parameters
Lines connect data from one occasion, points show measured concentrations, the blue horizontal line indicated time of dosing. 

```{r, fig.height=8.5, fig.width=11, warning=F, message=F}
# ------------------------------------------------------------------
#  Plot multiple dose data
# ------------------------------------------------------------------
# split by id.occ
nca_multiple_dose_split <- ind_data_split(nca_multiple_doses, 
                                          n_per_page = 12, id="ID.OCC")
nca_multiple_conc_split <- ind_data_split(nca_multiple_concs,
                                          n_per_page = 12, id="ID.OCC")

# IndividualPK profiles for multiple dose/ steady state
p1 <- vector("list", length(nca_multiple_conc_split))
for(i in 1:length(nca_multiple_conc_split)){
  p <-
    # concentrations
    ggplot(data=nca_multiple_conc_split[[i]], aes(x=TAPD, y=DV)) +
    geom_line() + geom_point() +
    # doses
    geom_vline(data = nca_multiple_dose_split[[i]], aes(xintercept=TAPD),
               col="blue", linetype="dashed") +
    facet_wrap(~ID.OCC, nrow=3, ncol=4,
               scales="free_y", labeller="label_both") +
    labs(x=labs_TAPD, y=labs_conc)
  p1[[i]] <- p
}

invisible(lapply(p1, grob_draw))
```

```{r, include=F}
if(params$print_results) {
  pdf(file=file.path(res_eda_dir, paste0("nca_multiple_dose_",delivery_date,".pdf")),
      height=8.5, width=11)
  invisible(lapply(p1, grob_draw))
  dev.off()
}
```

```{r, include=F}
# ------------------------------------------------------------------
# Merge conc and dose data into NCA object
# ------------------------------------------------------------------
nca_multiple_data <-
  PKNCAdata(nca_multiple_concs, nca_multiple_doses,
            formula.conc = DV~TAPD | ID.OCC,
            formula.dose = AMT~TAPD | ID.OCC,
            intervals = nca_intervals)

# ------------------------------------------------------------------
# Run NCA, extract and re-format results
# ------------------------------------------------------------------
nca_results_multiple <- pk.nca(nca_multiple_data)

# nca results
nca_results_multiple_extract <- nca_results_multiple$result
nca_results_multiple_extract <- 
  spread(nca_results_multiple_extract[, c("ID.OCC","PPTESTCD","PPORRES")],
         key = PPTESTCD, value = PPORRES)

# Merge back with covariate data (extract from dose object)
cov_extract_multiple <- 
  nca_results_multiple$data$dose$data[, c("ID.OCC", cols_cont_cov, base_cont_cov,
                                        cols_cat_cov, base_cat_cov,
                                        "OSID","OSTUDYID","AMT","FREQ", "DOSEFLAG")]

# unique(unique(covExtractMultiple$ID.OCC) == unique(nca_results_multiple_extract$ID.OCC))
results_multiple <- full_join(nca_results_multiple_extract, 
                              cov_extract_multiple) %>% 
  rename(dose = AMT)

# ------------------------------------------------------------------
#  Merge restults from sinlge dose and multiple dose occasions
# ------------------------------------------------------------------
results <- rbind(results_single, results_multiple)

### Calculate dose normalized values
results$aucInfDoseNorm <- (results$aucinf.obs / results$dose)
results$aucLastDoseNorm <- (results$auclast / results$dose)
results$cmaxDoseNorm   <- (results$cmax / results$dose)
```

# Assessment of PK linearity, $AUC_{Inf}$ vs dose
## ANOVA

```{r, fig.height=5, fig.width=6.5, warning=F, message=F}
anova <- dose_proportionality_anova(results, x=dose, y=aucinf.obs, 
                                    plot.par = T)
anova[["plot"]] + 
  labs(y=expression(AUC[inf]~(mg/L%.%h)), 
       x=expression(Dose~(mg)))
```

## Linear model

```{r, fig.height=5, fig.width=6.5, warning=F, message=F}
linear <- dose_proportionality_linear(results, x=dose, y=aucinf.obs, 
                                      plot.par = T)
linear[["plot"]] + 
  labs(y=expression(AUC[inf]~(mg/L%.%h)), 
       x=expression(Dose~(mg)))
```

## Power-model
```{r, fig.height=5, fig.width=6.5, warning=F, message=F}
power <- dose_proportionality_power(results, x=dose, y=aucinf.obs, 
                                      plot.par = T)
power[["plot"]] + 
  labs(y=expression(AUC[inf]~(mg/L%.%h)), 
       x=expression(Dose~(mg)))
```

# Other graphics assessing PK linearity

```{r, fig.height=4.5, fig.width=5.5, warning=F, message=F}
# Dose should be numeric (it is set as a group within boxplot) using gg_continous_box
# That way the x-axis is still numeric and displayes the data as continous
# (important if large range of doses with long, not equally long steps between)
# str(results$dose)

linearity_plots_pool  <- vector("list", 0)

## Cmax ###
linearity_plots_pool$cmax <-
  gg_box_continuous(results, x=dose, y=cmax) + 
  labs(x=expression(Dose~(mg)), y=expression(C[max]~(mg/L)))

linearity_plots_pool$norm_cmax <-
  gg_box_continuous(results, x=dose, y=cmaxDoseNorm) + 
  geom_hline(aes(yintercept=mean(cmaxDoseNorm, na.rm = T)),
             linetype="dashed", size=0.6) +
  labs(y=expression(Dose~normalized~C[max]~(mg/L)),
       x = expression(Dose~(mg)))

## AUC Infinity ###
linearity_plots_pool$auc_inf <-
  gg_box_continuous(results, x=dose, y=aucinf.obs) + 
  labs(y=expression(AUC[inf]~(mg/L%.%h)), 
       x=expression(Dose~(mg)))

linearity_plots_pool$norm_auc_inf <-
  gg_box_continuous(results, x=dose, y=aucInfDoseNorm) + 
  geom_hline(aes(yintercept=mean(aucInfDoseNorm, na.rm = T)),
             linetype="dashed", size=0.6) +
  labs(y=expression(Dose~normalized~AUC[inf]~(mg/L%.%h)),
       x = expression(Dose~(mg)))

## AUC 24h ###
linearity_plots_pool$auc_last <-
  gg_box_continuous(results, x=dose, y=auclast) + 
  labs(y=expression(AUC[last]~(mg/L%.%h)), 
       x = expression(Dose~(mg)))

linearity_plots_pool$norm_auc_inf <-
  gg_box_continuous(results, x=dose, y=aucLastDoseNorm) + 
  geom_hline(aes(yintercept=mean(aucLastDoseNorm, na.rm = T)),
             linetype="dashed", size=0.6) +
  labs(y=expression(Dose~normalized~AUC[last]~(mg/L%.%h)), 
       x = expression(Dose~(mg)))

# t1/2 ##
linearity_plots_pool$t_half <-
  gg_box_continuous(results, x=dose, y=half.life) + 
  geom_hline(aes(yintercept=mean(half.life, na.rm = T)),
             linetype="dashed", size=0.6) +
  labs(y=expression("t"["1/2"]~(h)),
       x = expression(Dose~(mg)))

invisible(lapply(linearity_plots_pool, grob_draw))

if(params$print_results){
  pdf(file=file.path(res_eda_dir,
                     paste0("PK_linearity_assessment_pooled_", delivery_date,".pdf")),
      height=4.5, width=5.5)
  invisible(lapply(linearity_plots_pool , grob_draw))
  dev.off()
}
```

## Stratified by single/multiple dose

```{r, fig.height=5, fig.width=7.5, warning=F, message=F}
linearity_plots_split <- vector("list", 0)

## Cmax ###
linearity_plots_split$cmax_split <-
  gg_box_continuous(results, y=cmax, x=dose,
                  fill=DOSEFLAG, dodge=30) +
  labs(y=expression(C[max]~(mg/L)),
       x=expression(Dose~(mg)))

linearity_plots_split$norm_cmax_split <- 
  gg_box_continuous(results, y=cmaxDoseNorm, x=dose,
                  fill=DOSEFLAG, dodge=30) +
  # add grand mean
  geom_hline(aes(yintercept=mean(cmaxDoseNorm, na.rm = T)),
             linetype="dashed", size=0.6) +
  labs(y=expression(Dose~normalized~C[max]~(mg/L)),
       x=expression(Dose~(mg)))

## AUC Infinity ###
linearity_plots_split$auc_inf_split <-
  gg_box_continuous(results, y=aucinf.obs, x=dose,
                  fill=DOSEFLAG, dodge=30) +
  labs(y=expression(AUC[inf]~(mg/L%.%h)),
       x=expression(Dose~(mg)))

linearity_plots_split$norm_auc_inf_split <-
  gg_box_continuous(results, y=aucInfDoseNorm, x=dose,
                  fill=DOSEFLAG, dodge=30) +
  # add grand mean
  geom_hline(aes(yintercept=mean(auc_inf_dose_norm, na.rm = T)),
             linetype="dashed", size=0.6) +
  labs(y=expression(Dose~normalized~AUC[inf]~(mg/L%.%h)),
       x = expression(Dose~(mg)))

## AUC Last ###
linearity_plots_split$auc_inf_split <-
  gg_box_continuous(results, y=auclast, x=dose,
                  fill=DOSEFLAG, dodge=30) +
  labs(y=expression(AUC[last]~(mg/L%.%h)),
       x=expression(Dose~(mg)))

linearity_plots_split$norm_auc_inf_split <-
  gg_box_continuous(results, y=auclast, x=dose,
                  fill=DOSEFLAG, dodge=30) +
  # add grand mean
  geom_hline(aes(yintercept=mean(aucLastDoseNorm, na.rm = T)),
             linetype="dashed", size=0.6) +
  labs(y=expression(Dose~normalized~AUC[last]~(mg/L%.%h)),
       x = expression(Dose~(mg)))

## t1/2 ##
linearity_plots_split$t_half_split <-
  gg_box_continuous(results, y=half.life, x=dose,
                  fill=DOSEFLAG, dodge=30) +
  # add grand mean
  geom_hline(aes(yintercept=mean(half.life, na.rm = T)),
             linetype="dashed", size=0.6) +
  labs(y=expression("t"["1/2"]~(h)),
       x=expression(Dose~(mg)))

invisible(lapply(linearity_plots_split, grob_draw))

if(params$print_results){
  pdf(file=file.path(res_eda_dir,
                     paste0("PK_linearity_assessment_stratified_", delivery_date,".pdf")),
      height=8.5, width=11)
  invisible(lapply(linearity_plots_split, grob_draw))
  dev.off()
}
```

# Assessment of covariate effect on NCA parameters
## Continuous covariates
Add figure text here

```{r, fig.height=8, fig.width=10, warning=F, message=F}
dose_norm_cont <-
  ggduo(results, base_cont_cov, 
        c("aucInfDoseNorm","aucLastDoseNorm","cmaxDoseNorm"),
        types = list(continuous = "smooth_loess"),
        columnLabelsY = c("AUC[Inf]/Dose~(mg/L%.%h/mg)",
                          "AUC[last]/Dose~(mg/L%.%h/mg)",
                          "C[max]/Dose~(mg/L/mg)"),
        labeller="label_parsed")

dose_norm_cont

if(params$print_results){
  pdf(file=file.path(res_eda_dir,
                     paste0("NCA_cont_covariates", delivery_date,".pdf")),
      height=8, width=10)
  print(dose_norm_cont)
  dev.off()
}
```

## Categorical covariates

```{r, fig.height=8, fig.width=10, warning=F, message=F}
dose_norm_cat <-
  ggduo(results, base_cat_cov,
        c("aucLastDoseNorm","aucLastDoseNorm","cmaxDoseNorm"),
        types = list(comboVertical = ally_box),
        columnLabelsY = c("AUC[Inf]/Dose~(mg/L%.%h/mg)",
                          "AUC[Last]/Dose~(mg/L%.%h/mg)",
                          "C[max]/Dose~(mg/L/mg)"),
        labeller="label_parsed")

dose_norm_cat

if(params$print_results){
  pdf(file=file.path(res_eda_dir,
                 paste0("NCA_cat_covariates", delivery_date,".pdf")),
      height=8, width=10)
  print(dose_norm_cat)
  dev.off()
}
```

## Stratified by study
```{r, fig.height=8, fig.width=5, warning=F, message=F}
dose_norm_study <-
  ggduo(results,"OSTUDYID", 
        c("aucLastDoseNorm","aucLastDoseNorm","cmaxDoseNorm"),
        types = list(comboVertical = ally_box),
        columnLabelsY = c("AUC[Inf]/Dose~(mg/L%.%h/mg)",
                          "AUC[Last]/Dose~(mg/L%.%h/mg)",
                          "C[max]/Dose~(mg/L/mg)"),
        labeller="label_parsed")
dose_norm_study

if(params$print_results){
  pdf(file=file.path(res_eda_dir,
                 paste0("NCA_by_study", delivery_date,".pdf")),
      height=8, width=5)
  print(dose_norm_study)
  dev.off()
}
```
